app: replace the handmade Roby plugin mechanism by the widely used Gem-based convention
Plugins must now place a file that is called roby-<plugin_name> in
either RUBYLIB (for non-gem packages), or in the Gem's load path.

There is no load vs. enable separation anymore. The registration was
really there *because* Roby allowed to load a file that had no
relationship with the plugin name. That's not the case anymore, so
loading the plugin file *is* enabling it.