interface: delay sending the exceptions until push_pending_notifications
The job notifications are delayed until the end of a cycle, but the
exceptions were sent right away. This leads in some cases to receiving
the job information after the exception, ignoring the exception.

Queue the exceptions until the end of the cycle as well, to fix ordering.
It still assumes that the receiving end will stop its listeners on
FINALIZE *after* the exceptions have been processed, something that
cabb633ac37f1f471996daee08be375af90e9c28 implemented for Async::Interface