core: improve handling of delayed arguments during merge
The current policy was "all or nothing or everything". Either
1. the argument was set to an actual value, and it has to stay
   at this value forever
2. the argument was set to a delayed value, and it could be
   updated by anything
3. the argument was not set, and it could be set to anything

This was a bit of a ... limiting policy for the purpose of merging
plans (i.e. in Syskit). The functionality for which this was designed
(default arguments) worked just fine, a (1) and (2) are good policies
for it. For referenced arguments (the from(...) syntax), it leads to
merging things that ultimately point to completely different arguments.

To deal with the difference in semantic between the two types of
delayed arguments, and provide a framework to more flexibly handle
merging arguments in a plan merge system, this commit adds:
- the #strong? predicate. strong? separates the arguments are
  really meant to be arguments (true) with the ones are meant to be
  used as defaults (i.e. can be overriden, false)
- the #can_semantic_merge? predicate and #semantic_merge! operation
  that allow computing a merged argument, resulting of two existing
  arguments. I hope this can be used down the line to provide
  interesting merging range semantic in plans.